requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
psycopg2-binary
jinja2
aiofiles
deep-translator
cohere
